requirements.txt
KS-BOT-Client-Python